python/requirements.txt
python-magic==0.4.27
google-cloud-storage==3.1.0
google-cloud-spanner==3.55.0
fastapi==0.115.12
uvicorn==0.34.3
asyncer==0.0.8
beautifulsoup4==4.13.4
google-cloud-pubsub==2.30.0
csr-va-common-metrics==0.2.6
google-cloud-bigquery==3.36.0
